V17 [Patch]

+++ ScheduledJobsWebApp/Lib/XTI_Jobs/EventMonitor.cs
@@ -4,81 +4,43 @@ public sealed class EventMonitor
 {
     private readonly IJobDb db;
     private readonly IJobActionFactory jobActionFactory;
+    private readonly ITransformedEventData transformedEventData;
     private readonly EventKey eventKey;
     private readonly JobKey jobKey;
     private readonly DateTimeOffset eventRaisedStartTime;
 
-    internal EventMonitor(IJobDb db, IJobActionFactory jobActionFactory, EventKey eventKey, JobKey jobKey, DateTimeOffset eventRaisedStartTime)
+    internal EventMonitor(IJobDb db, IJobActionFactory jobActionFactory, ITransformedEventData transformedEventData, EventKey eventKey, JobKey jobKey, DateTimeOffset eventRaisedStartTime)
     {
         this.db = db;
         this.jobActionFactory = jobActionFactory;
+        this.transformedEventData = transformedEventData;
         this.eventKey = eventKey;
         this.jobKey = jobKey;
         this.eventRaisedStartTime = eventRaisedStartTime;
     }
 
-    public async Task Run(CancellationToken stoppingToken)
+    public async Task<TriggeredJob[]> Run(CancellationToken stoppingToken)
     {
+        if (eventKey.IsOnDemand())
+        {
+            throw new ArgumentException($"Unable to monitor on demand job '{eventKey.DisplayText}'");
+        }
         await db.DeleteJobsWithNoTasks(eventKey, jobKey);
         var retryJobs = await db.RetryJobs(eventKey, jobKey);
+        var jobRunner = new JobRunner(db, jobActionFactory);
+        var triggeredJobs = new List<TriggeredJob>();
         foreach(var retryJob in retryJobs)
         {
-            var triggeredJob = new TriggeredJob(db, retryJob);
-            var nextTask = await triggeredJob.StartNextTask();
-            while (nextTask != null)
-            {
-                nextTask = await ExecuteTask(stoppingToken, triggeredJob, nextTask);
-            }
+            var triggeredJob = await jobRunner.StartRetry(retryJob, stoppingToken);
+            triggeredJobs.Add(triggeredJob);
         }
         var pendingJobs = await db.TriggerJobs(eventKey, jobKey, eventRaisedStartTime);
         foreach (var pendingJob in pendingJobs)
         {
-            var taskData = await jobActionFactory.TransformSourceData(pendingJob.SourceKey, pendingJob.SourceData);
-            var firstTasks = jobActionFactory.FirstTasks(taskData);
-            var triggeredJob = new TriggeredJob(db, pendingJob);
-            var nextTask = await triggeredJob.Start(firstTasks);
-            while (nextTask != null)
-            {
-                nextTask = await ExecuteTask(stoppingToken, triggeredJob, nextTask);
-            }
+            var taskData = await transformedEventData.TransformEventData(pendingJob.SourceKey, pendingJob.SourceData);
+            var triggeredJob = await jobRunner.StartNew(pendingJob, taskData, stoppingToken);
+            triggeredJobs.Add(triggeredJob);
         }
-    }
-
-    private async Task<TriggeredJobTask?> ExecuteTask(CancellationToken stoppingToken, TriggeredJob triggeredJob, TriggeredJobTask currentTask)
-    {
-        TriggeredJobTask? nextTask;
-        var jobAction = jobActionFactory.CreateJobAction(currentTask);
-        JobActionResult result;
-        try
-        {
-            result = await jobAction.Execute(stoppingToken);
-            await currentTask.Completed(result.PreserveData, result.NextTasks);
-            nextTask = await triggeredJob.StartNextTask();
-        }
-        catch(CancelJobException cancelJobEx)
-        {
-            await currentTask.CancelJob(cancelJobEx.Reason);
-            nextTask = null;
-        }
-        catch (Exception ex)
-        {
-            JobErrorResult errorResult;
-            try
-            {
-                errorResult = await jobAction.OnError(ex);
-            }
-            catch
-            {
-                errorResult = new JobErrorResultBuilder(currentTask.Model).Build();
-            }
-            nextTask = await currentTask.Failed
-            (
-                errorResult.UpdatedStatus, 
-                errorResult.RetryAfter, 
-                errorResult.NextTasks, 
-                ex
-            );
-        }
-        return nextTask;
+        return triggeredJobs.ToArray();
     }
 }
